Linear texture fix
Removed a few pixels in the bottom left of Sepiroth's DMW icon. This fixes an issue where, when using linear texture filtering, boxes are rendered behind damage numbers. (this issue is preset with or without the hd texture pack, but the fix causes no issues)
Also fixed a cropping issue where Sephiroth's hair went into the bottom of Chocobo's icon.